Refactor Agents to Separate File
